feat(staking): drop "rewards" (mislabeled withdrawals) from activity, merge transformers
