+++ apps/browser-extension-wallet/src/views/browser-view/features/activity/helpers/index.ts
@@ -2,3 +2,4 @@ export * from './pending-tx-transformer';
 export * from './tx-history-transformer';
 export * from './is-tx-with-assets';
 export * from './filter-outputs-by-tx-direction';
+export * from './types';
