+++ apps/browser-extension-wallet/src/views/browser-view/features/activity/helpers/types.ts
@@ -1,0 +1,3 @@
+import type { AssetActivityItemProps } from '@lace/core';
+
+export type TransformedTx = Omit<AssetActivityItemProps, 'onClick'>;
